tests/input-device: Add artificial test data for deriving type from evdev properties
If a device is positively identified as an accelerometer, pointing stick
or clickpad, then we don't need to second-guess it.

In practice this does not change the result for any device in our
test data, so add some artificial records that exercise this.

Origin: https://github.com/libsdl-org/SDL/pull/7837
Signed-off-by: Simon McVittie <smcv@collabora.com>