tests/input-device: Add X-Box One S controller via xpadneo 0.9.x
This is an out-of-tree driver providing an alternative interface to
X-Box gamepads. Compared with the in-kernel driver, it adds four
pseudo-buttons used to indicate which profile is active, and an
optional "misc" axis that combines the two triggers into a rudder.

Thanks: @kakra
Origin: https://github.com/libsdl-org/SDL/pull/10076
Signed-off-by: Simon McVittie <smcv@collabora.com>