srt-logger --mkfifo: New CLI tool entry point
If we could rely on `mktemp -d` being fully portable, then we could
do this from shell script, but we probably can't rely on that.

Similarly, we probably can't rely on `$XDG_RUNTIME_DIR` being set, or
`$TMPDIR` being a filesystem that is able to hold fifo objects.

This has the same statically-linked dependencies as srt-logger, so make
it part of the srt-logger binary to avoid wasting space.

Signed-off-by: Simon McVittie <smcv@collabora.com>