logger-0.bash: Add a convenience bash binding for srt-logger
There are a few places where we want to log the output of a shell
script without having to execute srt-logger wrapped around a second
shell script as a subprocess. steam.sh and steamwebhelper.sh already
use srt-logger, and bin_steam.sh might follow soon.

To achieve this, we need to use either a bash coprocess or at least
one fifo, because process substitution only creates a one-way pipe.
Currently all the scripts where we want to use the logger are bash
scripts (as opposed to pure POSIX sh) so we can still take advantage
of other bashisms: process substitution and named redirections.

Even when using those bash features, the synchronization required
between the logger and its parent is non-trivial, so putting it in a
shell library makes it easier to achieve.

Signed-off-by: Simon McVittie <smcv@collabora.com>