Revert "pv-runtime: Add a mechanism to take custom CA certificates from the host"
This was implemented in a way that assumed it would only be enabled
on specially-flagged developer systems, making it acceptable for the
implementation to be non-robust and make assumptions about the host
systems on which it would be enabled; but current Steam betas enable
it for the steamwebhelper unconditionally, which would cause serious
regressions on non-Debian-derived, non-Arch-derived host OSs.

Revert the addition of this feature to avoid regressions. Before
reinstating it, we will either need to arrange for Steam to enable it
conditionally as previously planned, or make the implementation fail
gracefully and harmlessly on non-Debian-compatible host systems.

This reverts commit d6a519f5590462892233a443b8628acb452cb713.

steamrt/tasks#416

Signed-off-by: Simon McVittie <smcv@collabora.com>