pv-runtime: Don't warn if library aliases are missing from secondary archs
Our older runtimes are symmetrical, so the vast majority of libraries
(and in particular, the ones that need aliases for an alternate SONAME)
are available for both x86_64 and i386.

However, in some of our newer runtimes like medic, we've started
dropping i386 libraries that don't strictly need to exist in order to
save some space. This means we can't necessarily assume that libraries
that need aliases will exist for every architecture, so we shouldn't
emit warnings if they turn out not to exist.

steamrt/tasks#367

Signed-off-by: Simon McVittie <smcv@collabora.com>