logger: Add an early call to g_ascii_strtoull() to avoid a GLib bug
There is a difficult-to-reproduce failure mode in which
g_ascii_string_to_unsigned() can fail to parse a valid integer, which
was causing our tests to fail (but only on i386 and only in a specific
non-interactive build environment). As far as I can make out, this is
an error-handling issue in the way g_ascii_string_to_unsigned() calls
g_ascii_strtoull(): if newlocale() clobbers errno (which it is within
its rights to do, according to the usual errno conventions), then
g_ascii_strtoull() would interpret that as a strtoull_l() failure.

We can work around this by ensuring that our first call to
g_ascii_strtoull() is one that doesn't matter, so that by the time we
actually call g_ascii_string_to_unsigned(), the cached C locale object
inside GLib has already been initialized.

steamrt/tasks#481

Signed-off-by: Simon McVittie <smcv@collabora.com>