capture-libs: Add a new flag to suppress warnings
Sometimes we find ourselves capturing a library for which the
dependencies can legitimately be missing, for example
`libnvidia-pkcs11.so.*` on Ubuntu 24.04 (which depends on OpenSSL 1.1,
which Ubuntu no longer provides).

This is really an OS bug (providing a library without its dependencies),
and ideally Ubuntu should stop providing `libnvidia-pkcs11.so.*`
(like Debian did), but it results in confusing messages that could make
a user wrongly believe that it's the root cause of a genuine problem.

steamrt/tasks#488

Signed-off-by: Simon McVittie <smcv@collabora.com>