capture-libs: Add a --level-prefix option
This will allow the output of capsule-capture-libs (and in principle
other tools) to be piped to a log handler such as `logger --prio-prefix`,
`systemd-cat --level-prefix=true` or `srt-logger --parse-level-prefix`
which will filter or format the messages according to their severity.

As currently implemented, fatal errors are emitted as LOG_ERR, warnings
are LOG_WARNING, and all debug messages are LOG_DEBUG. A facility such
as LOG_USER is not included: this is optional for logger(1), and not
allowed for systemd-cat(1) and srt-logger.

steamrt/tasks#444

Signed-off-by: Simon McVittie <smcv@collabora.com>