pv-runtime: Silence spurious warnings for libnvidia-pkcs11 on Ubuntu 24.04
My understanding is that these libraries are dlopen'd by other parts
of the Nvidia driver for confidential-computing GPGPU use-cases, and
unlikely to be critical for gaming.

There are currently two implementations, one that depends on OpenSSL 1.1
and one that depends on OpenSSL 3.0. The way they are packaged in Ubuntu
means that in 24.04, the one for OpenSSL 1.1 is present but does not
have its dependencies satisfied, leading to confusing warning messages
from capsule-capture-libs that could confuse our users. Silence those
messages by using the new "quiet:" pattern prefix.

(The message is actually reduced to LOG_DEBUG level rather than being
silenced, so CAPSULE_DEBUG=tool will reinstate it if someone needs to
look at capsule-capture-libs behaviour in fine detail.)

steamrt/tasks#488

Signed-off-by: Simon McVittie <smcv@collabora.com>