launch-options: Switch into correct LDLP runtime before launch adverbs
It's reasonable for steam-launch-wrapper to want to adjust the PATH
or LD_LIBRARY_PATH that were provided by the LDLP runtime, for example
for steamrt/tasks#489. That won't work if we immediately overwrite them
by switching into a different LDLP runtime.

This also means that while testing a non-default LDLP runtime, we'll
use the runtime under test (and not the LDLP runtime that Steam is
using, if different) to run the steam-launch-wrapper and reaper.

Signed-off-by: Simon McVittie <smcv@collabora.com>