Consistently disable GIO modules for executables with an RPATH
We'll be running these with the GLib that we located via the RPATH,
which (in general) will not be new enough to load GIO modules from the
host OS; so in every executable that is linked to GIO and has our RPATH,
we will want to turn off external modules.

This edits the global environment variables, so in programs that wrap
another command (such as steam-runtime-urlopen since steamrt/tasks#489),
we need to save the original environment first, and base the environment
of the wrapped command on the saved copy.

steamrt/tasks#505

Signed-off-by: Simon McVittie <smcv@collabora.com>