Avoid deprecated g_qsort_with_data()
For historical reasons g_qsort_with_data() "only" works with up to 2**31
items, so it won't necessarily work for pathologically large arrays
and therefore is deprecated.

The advantage of g_qsort_with_data() and its replacement g_sort_array()
is that GLib guarantees that they are a stable sort (will not permute
items that already compare equal), which is not a guarantee for glibc's
qsort() and qsort_r(). However, in each of these places we are sorting
the keys from a hash table, which are unique anyway, so there is no
advantage to a stable sort.

Signed-off-by: Simon McVittie <smcv@collabora.com>