pv-runtime: Generalize code to capture libraries related to glibc
Previously this was part of pv_runtime_collect_libc_family(), but in
fact it's equally applicable for other libraries.

For example, if we are using the host's libxkbcommon.so.0, we'll also
want the host's libxkbcommon-x11.so.0 if at all possible, because
libxkbcommon-x11 uses private data structures from libxkbcommon and
will crash if they are not sufficiently similar.

steamrt/tasks#530

Signed-off-by: Simon McVittie <smcv@collabora.com>