pv-runtime: If using host libxkbcommon, try to use host libxkbcommon-x11
libxkbcommon is not really a graphics driver dependency as such, but
it can end up being pulled in by graphics stack components, for example
MangoHUD 0.7.2.

libxkbcommon-x11 looks into private data structures from libxkbcommon,
which is a reasonable thing to do since they come from the same source
code and are upgraded at the same time, but can cause crashes if we
pick up one but not the other from the graphics stack provider; so if
we pick up libxkbcommon as a dependency of MangoHUD, we also need to
take the matching libxkbcommon-x11 if it exists.

If the graphics stack provider has libxkbcommon but not libxkbcommon-x11,
there's really nothing we can do about that, so we will have to combine
this change with trying harder to get Steam client packaging to pull in
libxkb-common-x11.so.0.

steamrt/tasks#530

Signed-off-by: Simon McVittie <smcv@collabora.com>