library-cmp: When listing symbols, ignore the name of each verdef
Conventionally, for each version definition in a shared library, there
is a symbol named after the verdef with absolute value 0. For example,
libdbus has a symbol LIBDBUS_1_3@LIBDBUS_1_3, and similarly libxkbcommon
has symbols V_0.8.0@V_0.8.0 and so on.

These will typically not match any of the glob patterns that we use to
mark symbols as being public or private, leading to spurious warnings
like this one:

    i386-linux-gnu-capsule-capture-libs: warning: we are assuming "V_0.5.0" to be private, but it's just a guess

So far, we have avoided this because we have not used the "symbols"
comparison order for any library that has verdefs, but I'm intending
to use "versions,name,symbols" for libxkbcommon.

Signed-off-by: Simon McVittie <smcv@collabora.com>