logger: Make the "Log opened" message consistent with later messages
The other messages have a timestamp in square brackets, so let's
match that here. One difference is that when recording the log as having
been opened, we mention the time zone: to reduce redundancy, we don't
write this on subsequent lines.

The test needs adjusting slightly so that it can still skip the
"Log opened" line.

steamrt/tasks#537

Signed-off-by: Simon McVittie <smcv@collabora.com>