dialog: Prefer to look for dialog-ui in $STEAM_RUNTIME_SCOUT
Historically $STEAM_RUNTIME has been both an input to steam.sh
(switching to a different LDLP runtime for testing, or disabling it
altogether) and an output from steam.sh/run.sh (indicating that we are
now inside the LDLP runtime), and when used as an input it can be set
to either a boolean or an absolute path.

This is needlessly confusing, and we're trying to phase out at least
its role as a path input, replacing it with $STEAM_RUNTIME_SCOUT which
is already used by the "Steam Linux Runtime 1.0 (scout)" hybrid
container runtime.

steamrt/tasks#520

Signed-off-by: Simon McVittie <smcv@collabora.com>