populate-depot: Set filter_exclusive_priority equal to major version
Previously we were setting `filter_exclusive_priority` to 1 for all
"pure" (non-layered) container runtimes as per steamrt/tasks#426,
but in fact that was an oversight. Instead, we need to set a unique
priority for each distinct ABI (scout, soldier, sniper and so on).

This is because `Properties → Compatibility → Force the use of…`
lists all compatibility tools that have a `filter_exclusive_priority`
equal to the one that has been set on the app/game. We don't want
to list SLR 2.0 (soldier) for games that require the
Steam Runtime 3 (sniper) ABI, and we don't want to list a possible
future SLR 4 (medic) for sniper games either.

We assume here that future runtimes will follow the pattern set by
steamrt5 and have boring, pragmatic codenames. If we reach steamrt9
(which we should start work on in around 2031, assuming we continue
to do one Steam Runtime per Debian release) without the design having
changed, then we will need to do some renumbering at that point.

steamrt/tasks#597

Signed-off-by: Simon McVittie <smcv@collabora.com>