Add support for OpenXR manifests
OpenXR manifests are similar to Vulkan ICDs: JSON files with library
path or soname.
The main difference is that there is only one active runtime for a given
ABI.
OpenXR loader doesn't allow overriding the search path, but only setting
the explicit manifest file, which must work for the application ABI.

In this commit, we assume that there is a single manifest, which is how
most runtimes work.

Signed-off-by: Patrick Nicolas <patricknicolas@laposte.net>